requirements.txt
fastapi
motor
pyyaml
pydantic-settings
uvicorn
jinja2